Refs #265. Renamed directory and file name and fixed some issues during reduciton test.
